+++ scripts/demos/haply_teleoperation.py
@@ -327,7 +327,7 @@ def run_simulator(
         left_finger_forces = left_finger_sensor.data.net_forces_w[0, 0]
         right_finger_forces = right_finger_sensor.data.net_forces_w[0, 0]
         total_contact_force = (left_finger_forces + right_finger_forces) * 0.5
-        haply_device.push_force(total_contact_force.unsqueeze(0))
+        haply_device.push_force(forces=total_contact_force.unsqueeze(0), position=torch.tensor([0]))
 
 
 def main():